Fleshes out the readme and updates the demo.
The demo was out of date relative to tests, so I brought it in line with tests (basically a mirror of one of the tests but cleaned up for easy reading).

Readme now should allow me to use this library or dev on this library in 6 months when I have forgotten everything.  Also includes mention of how to use the adapter.

Adds a note to the MPT verifier that it doesn't work with unset/0 values.