Adds support for `ethers` JsonRpcProvider.
This mechanism for detecting an ethers JsonRpcProvider is a bit sketchy, but ethers is a popular enough library that it is worth it for now.  The next version of ethers should have a bridge that may work which would be the cleanere way to handle this, though TBD if it will uspport `eth_getProof`.

Also fixes a bug in `getblockByNumberFactory` where if you pass in `latest` it would try to fetch `0xlatest`.